Added save button to the action bar; Action bar title changes according to current fragment in view; Nav Host Fragment added to main activity
