+++ android-fmm/app/src/main/java/com/example/androidfmm/alarm/AlarmListFragment.kt
@@ -1,12 +1,12 @@
 package com.example.androidfmm.alarm
 
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
+import android.util.Log
+import android.view.*
 import androidx.cardview.widget.CardView
 import androidx.fragment.app.Fragment
 import androidx.databinding.DataBindingUtil
+import androidx.navigation.Navigation
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.androidfmm.R
 import com.example.androidfmm.databinding.FragmentAlarmListBinding
@@ -15,7 +15,7 @@ import kotlinx.android.synthetic.main.fragment_alarm_list.*
 class AlarmListFragment : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        retainInstance = true
+        setHasOptionsMenu(true)
     }
 
     override fun onCreateView(
@@ -26,15 +26,12 @@ class AlarmListFragment : Fragment() {
         val binding = DataBindingUtil.inflate<FragmentAlarmListBinding>(
             inflater, R.layout.fragment_alarm_list, container, false)
         binding.alarmList = this
-        
+
         return binding.root
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-
-        //Test method call
-//        val testList = generateTest(100)
 
         val alarmItemsList = listOf(
             AlarmItem(0, dayOfWeek = "Monday", alarmTime = "12:00PM", alarmDate = "8/31/20"),
@@ -53,19 +50,20 @@ class AlarmListFragment : Fragment() {
         alarm_list_fragment.setHasFixedSize(true)
     }
 
-    fun goToAlarmCreate() {
-
+    override fun onCreateOptionsMenu(menu: Menu, menuInflater: MenuInflater) {
+        menuInflater.inflate(R.menu.actionbar_menu, menu)
+        super.onCreateOptionsMenu(menu, menuInflater)
     }
 
-    // Test method
-//    private fun generateTest(size: Int): List<AlarmItem> {
-//        val list = ArrayList<AlarmItem>()
-//
-//        for (i in 0 until size) {
-//            val item = AlarmItem(i.toString())
-//            list += item
-//        }
-//
-//        return list
-//    }
+    override fun onOptionsItemSelected(item: MenuItem): Boolean = when (item.itemId) {
+        R.id.nav_create_new_alarm -> {
+            Navigation.findNavController(requireActivity(), R.id.nav_host_fragment).navigate(R.id.createAlarmFragment)
+            Log.i("MENU", "MENU NAV CLICKED")
+            true
+        }
+        else -> {
+            Log.i("MENU", "WAT")
+            super.onOptionsItemSelected(item)
+        }
+    }
 }
