+++ android-fmm/app/src/main/java/com/example/androidfmm/createalarm/CreateAlarmFragment.kt
@@ -1,9 +1,8 @@
 package com.example.androidfmm.createalarm
 
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
+import android.util.Log
+import android.view.*
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import com.example.androidfmm.R
@@ -12,6 +11,7 @@ import com.example.androidfmm.databinding.FragmentCreateAlarmBinding
 class CreateAlarmFragment: Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        setHasOptionsMenu(true)
     }
 
     override fun onCreateView(
@@ -24,5 +24,21 @@ class CreateAlarmFragment: Fragment() {
         binding.createAlarm = this
 
         return binding.root
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu, menuInflater: MenuInflater) {
+        menuInflater.inflate(R.menu.actionbar_save, menu)
+        super.onCreateOptionsMenu(menu, menuInflater)
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean = when (item.itemId) {
+        R.id.save_button -> {
+            Log.i("MENU", "SAVE BUTTON")
+            true
+        }
+        else -> {
+            Log.i("MENU", "WAT_CREATE")
+            super.onOptionsItemSelected(item)
+        }
     }
 }
