+++ android-fmm/app/src/main/res/values/strings.xml
@@ -4,4 +4,5 @@
     <string name="alarm_name_placeholder">Alarm Name</string>
     <string name="save_button">Save</string>
     <string name="new_alarm">+</string>
+    <string name="save_button_toolbar">Save</string>
 </resources>
